Merge main to stable

+++ jenkins/cicd/cortx-ha-cicd.sh
@@ -19,21 +19,48 @@
 set -x
 
 BASE_DIR=$(realpath "$(dirname $0)/../../")
+HA1="1"
+HA2="2"
+
+usage() {
+    echo """
+    $ ./cortx-ha-cicd.sh [-v <cortx-ha major version>]
+"""
+}
+
+while getopts ":v:h" o; do
+    case "${o}" in
+        v)
+            VERSION=${OPTARG}
+            ;;
+        h)
+            usage
+            ;;
+        *)
+            usage
+            ;;
+    esac
+done
+
+[ -z "$VERSION" ] && VERSION="${HA2}"
 
 # Copy Backend files
 TMPHA="${BASE_DIR}"/dist/tmp/cortx/ha
-rm -rf "${TMPHA}"
-mkdir -p "${TMPHA}"
-cp -rs "$BASE_DIR"/ha/* "${TMPHA}"
 
-mkdir -p /etc/cortx/ha/
-ls "${BASE_DIR}"/jenkins/cicd/etc/
-cp -rf "${BASE_DIR}"/jenkins/cicd/etc/* /etc/cortx/ha/
+if [ "$VERSION" == "${HA1}" ]
+then
+    rm -rf "${TMPHA}"
+    mkdir -p "${TMPHA}"
+    cp -rs "$BASE_DIR"/ha/* "${TMPHA}"
 
-# Perform unit test
-python3 "${TMPHA}"/test/main.py "${TMPHA}"/test/unit
+    mkdir -p /etc/cortx/ha/ /var/log/seagate/cortx/ha
+    cp -rf "${BASE_DIR}"/jenkins/cicd/etc/* /etc/cortx/ha/
 
-#/usr/bin/cortxha node refresh --hard --data-only || exit 1
+    # Perform unit test
+    python3 "${TMPHA}"/test/main.py "${TMPHA}"/test/unit
 
-/usr/lib/ocf/resource.d/seagate/hw_comp_ra meta-data
-/usr/lib/ocf/resource.d/seagate/iem_comp_ra meta-data
+    /usr/lib/ocf/resource.d/seagate/hw_comp_ra meta-data
+else
+    echo "no unit tests added for HA > 2 version yet."
+fi
+
